requirements.txt
fastapi
uvicorn
xmltodict
psycopg2-binary